feat: add public authentication classes with registration functionality
Used clean architecture, the delivery mechanism is the authentication router.
Data will flow from the route into the PublicAuthenticationController method.
The method will create an request model for the use-case. This model will flow
through the InputPort interface method into the use-case. The use-case will
perfome some logic and validation using value-objects etc. It will form an
response model and push it throug the OutputPort interface methode to the presenter.
The route will get the response from the presenter and send it to the client.