feat: add functions, timeouts and error messages to increase security
Added timeouts to get the same response timing for failed and successful responses.
Changed response success to true if an user doesnt exist in the db to avoid collection
of valid and used emails by an attacker.
Added some dummy methods if an user doesnt exist in the db.
Changed some notification emails from await to fire and forget.